+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Support for CSS customization through the use of CSS handles.
+
 ## [0.13.2] - 2019-12-18
 
 ### Added